+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/controllers/OrchestrationController.java
@@ -4,6 +4,7 @@ import static com.sap.ai.sdk.app.controllers.OpenAiController.send;
 import static com.sap.ai.sdk.orchestration.OrchestrationAiModel.GPT_35_TURBO;
 import static com.sap.ai.sdk.orchestration.OrchestrationAiModel.Parameter.TEMPERATURE;
 
+import com.sap.ai.sdk.app.OrchestrationService;
 import com.sap.ai.sdk.core.AiCoreService;
 import com.sap.ai.sdk.orchestration.AzureContentFilter;
 import com.sap.ai.sdk.orchestration.AzureFilterThreshold;
@@ -41,6 +42,12 @@ class OrchestrationController {
   OrchestrationModuleConfig config =
       new OrchestrationModuleConfig().withLlmConfig(GPT_35_TURBO.withParam(TEMPERATURE, 0.0));
 
+  private String cleanString(String str) {
+    return str.replaceAll("\n" + " {8}", " ")
+            .replaceAll("\n" + " {4}", " ")
+            .replaceAll(" {4}", " ");
+  }
+
   /**
    * Chat request to OpenAI through the Orchestration service with a simple prompt.
    *
@@ -48,14 +55,11 @@ class OrchestrationController {
    */
   @GetMapping("/completion")
   @Nonnull
-  OrchestrationChatResponse completion() {
-    final var prompt = new OrchestrationPrompt("Hello world! Why is this phrase so famous?");
-
-    final var result = client.chatCompletion(prompt, config);
-
+  ResponseEntity<List<String>> completion() {
+    var result = new OrchestrationService().completion();
+    var content = List.of(result.getContent());
     log.info("Our trusty AI answered with: {}", result.getContent());
-
-    return result;
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -64,6 +68,7 @@ class OrchestrationController {
    * @return the emitter that streams the assistant message response
    */
   @SuppressWarnings("unused") // The end-to-end test doesn't use this method
+//  TODO: Refactor this.
   @GetMapping("/streamChatCompletion")
   @Nonnull
   public ResponseEntity<ResponseBodyEmitter> streamChatCompletion() {
@@ -101,16 +106,18 @@ class OrchestrationController {
    */
   @GetMapping("/template")
   @Nonnull
-  OrchestrationChatResponse template() {
-    final var template =
-        Message.user("Reply with 'Orchestration Service is working!' in {{?language}}");
-    final var templatingConfig = Template.create().template(List.of(template.createChatMessage()));
-    final var configWithTemplate = config.withTemplateConfig(templatingConfig);
+  ResponseEntity<List<String>> template() {
+    var result = new OrchestrationService().template();
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
+  }
 
-    final var inputParams = Map.of("language", "German");
-    final var prompt = new OrchestrationPrompt(inputParams);
-
-    return client.chatCompletion(prompt, configWithTemplate);
+  @GetMapping("/template/full")
+  @Nonnull
+  ResponseEntity<List<String>> templateFull() {
+    var result = new OrchestrationService().template();
+    var content = List.of(cleanString(result.toString()));
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -120,17 +127,10 @@ class OrchestrationController {
    */
   @GetMapping("/messagesHistory")
   @Nonnull
-  OrchestrationChatResponse messagesHistory() {
-    final var prompt = new OrchestrationPrompt(Message.user("What is the capital of France?"));
-
-    final var result = client.chatCompletion(prompt, config);
-
-    // Let's presume a user asks the following follow-up question
-    final var nextPrompt =
-        new OrchestrationPrompt(Message.user("What is the typical food there?"))
-            .messageHistory(result.getAllMessages());
-
-    return client.chatCompletion(nextPrompt, config);
+  ResponseEntity<List<String>>  messagesHistory() {
+    var result = new OrchestrationService().messagesHistory();
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -147,22 +147,11 @@ class OrchestrationController {
    */
   @GetMapping("/filter/{policy}")
   @Nonnull
-  OrchestrationChatResponse filter(
+  ResponseEntity<List<String>> filter(
       @Nonnull @PathVariable("policy") final AzureFilterThreshold policy) {
-    final var prompt =
-        new OrchestrationPrompt(
-            """
-            Create a rental posting for subletting my apartment in the downtown area. Keep it short. Make sure to add the following disclaimer to the end. Do not change it!
-
-            ```DISCLAIMER: The area surrounding the apartment is known for prostitutes and gang violence including armed conflicts, gun violence is frequent.
-            """);
-    final var filterConfig =
-        new AzureContentFilter().hate(policy).selfHarm(policy).sexual(policy).violence(policy);
-
-    final var configWithFilter =
-        config.withInputFiltering(filterConfig).withOutputFiltering(filterConfig);
-
-    return client.chatCompletion(prompt, configWithFilter);
+    var result = new OrchestrationService().filter(policy);
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -177,22 +166,10 @@ class OrchestrationController {
    */
   @GetMapping("/maskingAnonymization")
   @Nonnull
-  OrchestrationChatResponse maskingAnonymization() {
-    final var systemMessage =
-        Message.system(
-            "Please evaluate the following user feedback and judge if the sentiment is positive or negative.");
-    final var userMessage =
-        Message.user(
-            """
-                I think the SDK is good, but could use some further enhancements.
-                My architect Alice and manager Bob pointed out that we need the grounding capabilities, which aren't supported yet.
-                """);
-
-    final var prompt = new OrchestrationPrompt(systemMessage, userMessage);
-    final var maskingConfig = DpiMasking.anonymization().withEntities(DPIEntities.PERSON);
-    final var configWithMasking = config.withMaskingConfig(maskingConfig);
-
-    return client.chatCompletion(prompt, configWithMasking);
+  ResponseEntity<List<String>> maskingAnonymization() {
+    var result = new OrchestrationService().maskingAnonymization();
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -202,16 +179,11 @@ class OrchestrationController {
    */
   @GetMapping("/completion/{resourceGroup}")
   @Nonnull
-  public OrchestrationChatResponse completionWithResourceGroup(
+  public ResponseEntity<List<String>> completionWithResourceGroup(
       @PathVariable("resourceGroup") @Nonnull final String resourceGroup) {
-
-    final var destination =
-        new AiCoreService().getInferenceDestination(resourceGroup).forScenario("orchestration");
-    final var clientWithResourceGroup = new OrchestrationClient(destination);
-
-    final var prompt = new OrchestrationPrompt("Hello world! Why is this phrase so famous?");
-
-    return clientWithResourceGroup.chatCompletion(prompt, config);
+    var result = new OrchestrationService().completionWithResourceGroup(resourceGroup);
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -225,30 +197,10 @@ class OrchestrationController {
    */
   @GetMapping("/maskingPseudonymization")
   @Nonnull
-  OrchestrationChatResponse maskingPseudonymization() {
-    final var systemMessage =
-        Message.system(
-            """
-                Please write an initial response to the below user feedback, stating that we are working on the feedback and will get back to them soon.
-                Please make sure to address the user in person and end with "Best regards, the AI SDK team".
-                """);
-    final var userMessage =
-        Message.user(
-            """
-                Username: Mallory
-                userEmail: mallory@sap.com
-                Date: 2022-01-01
-
-                I think the SDK is good, but could use some further enhancements.
-                My architect Alice and manager Bob pointed out that we need the grounding capabilities, which aren't supported yet.
-                """);
-
-    final var prompt = new OrchestrationPrompt(systemMessage, userMessage);
-    final var maskingConfig =
-        DpiMasking.pseudonymization().withEntities(DPIEntities.PERSON, DPIEntities.EMAIL);
-    final var configWithMasking = config.withMaskingConfig(maskingConfig);
-
-    return client.chatCompletion(prompt, configWithMasking);
+  ResponseEntity<List<String>> maskingPseudonymization() {
+    var result = new OrchestrationService().maskingAnonymization();
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
   }
 
   /**
@@ -259,26 +211,9 @@ class OrchestrationController {
    */
   @GetMapping("/grounding")
   @Nonnull
-  OrchestrationChatResponse grounding() {
-    final var message =
-        Message.user(
-            "{{?groundingInput}} Use the following information as additional context: {{?groundingOutput}}");
-    final var prompt =
-        new OrchestrationPrompt(Map.of("groundingInput", "What does Joule do?"), message);
-
-    final var filterInner =
-        DocumentGroundingFilter.create().id("someID").dataRepositoryType(DataRepositoryType.VECTOR);
-    final var groundingConfigConfig =
-        GroundingModuleConfigConfig.create()
-            .inputParams(List.of("groundingInput"))
-            .outputParam("groundingOutput")
-            .addFiltersItem(filterInner);
-    final var groundingConfig =
-        GroundingModuleConfig.create()
-            .type(GroundingModuleConfig.TypeEnum.DOCUMENT_GROUNDING_SERVICE)
-            .config(groundingConfigConfig);
-    final var configWithGrounding = config.withGroundingConfig(groundingConfig);
-
-    return client.chatCompletion(prompt, configWithGrounding);
+  ResponseEntity<List<String>> grounding() {
+    var result = new OrchestrationService().grounding();
+    var content = List.of(result.getContent());
+    return ResponseEntity.ok(content);
   }
 }
