+++ sample-code/spring-app/src/main/resources/static/index.html
@@ -69,6 +69,7 @@
             <li><a href="/orchestration/completion">/orchestration/completion</a></li>
             <li><a href="/orchestration/streamChatCompletion">/orchestration/streamChatCompletion</a></li>
             <li><a href="/orchestration/template">/orchestration/template</a></li>
+            <li><a href="/orchestration/template/full">/orchestration/template/full</a></li>
             <li><a href="/orchestration/messagesHistory">/orchestration/messagesHistory</a></li>
             <li><a href="/orchestration/filter/NUMBER_4">/orchestration/filter/NUMBER_4</a> Loose filter</li>
             <li><a href="/orchestration/filter/NUMBER_0">/orchestration/filter/NUMBER_0</a> Strict filter (fails)</li>
